server/requirements.txt
Flask
python-dotenv
flask-cors
